Rename file with usm tests.
