+++ dppl/tests/dppl_tests/__init__.py
@@ -1,2 +1,2 @@
 from .test_sycl_queue_manager import *
-from .test_sycl_memory_manager import *
+from .test_sycl_usm import *
